+++ src/orbital/lookup.rs
@@ -1,4 +1,4 @@
-pub const VERTEX_RELATIVE_POSITION: [[f64; 4]; 8] = [
+pub const VERTEX_RELATIVE_POSITION: [[f32; 4]; 8] = [
     [0.0, 0.0, 0.0, 0.0],
     [1.0, 0.0, 0.0, 0.0],
     [0.0, 1.0, 0.0, 0.0],
@@ -50,7 +50,7 @@ pub const _EDGE_MASKS: [i32; 256] = [
 ];
 
 // For each MC case, a list of triangles, specified as triples of edge indices, terminated by -1
-pub fn triangle_table(n: usize) -> Vec<i8> {
+pub fn triangle_table(n: usize) -> Vec<isize> {
     match n {
         0 => vec![-1],
         1 => vec![0, 3, 8, -1],
